requirements.txt
PyPDF2
openai
tqdm
colorama
python-dotenv
PyQt5
